requirements.txt
streamlit~=1.47.1
pillow~=11.3.0